feat: replaced common known drugs with ssp component
