--- packages/sections/src/common/KnownDrugs/index.js
@@ -1,1 +1,0 @@
-export { default as Body } from "./Body";
